+++ crates/vmi-arch-amd64/src/lib.rs
@@ -25,7 +25,7 @@ pub use self::{
     dr::{Dr0, Dr1, Dr2, Dr3, Dr6, Dr7},
     event::{
         EventCpuId, EventHypercall, EventInterrupt, EventIo, EventIoDirection, EventMemoryAccess,
-        EventMonitor, EventReason, EventSinglestep, EventWriteCr, MemoryAccessFlags,
+        EventMonitor, EventReason, EventSinglestep, EventWriteCr, EventWriteMsr, MemoryAccessFlags,
     },
     interrupt::{ExceptionVector, Idt, IdtAccess, IdtEntry, Interrupt, InterruptType},
     msr::{Msr, MsrEfer},